Trying to implement updating match progess
Every minute we make a get the json data and update the whole page
so we can update the score. Also every minute we calculate the percent
of the match that has passed and update the progress bar accordingly.
If we are in the half time break we do not update the progess bar nor
the remaining minutes. Depending on the current hour and the hour
that match should be played we either display the date or we display
the progress bar if the match is currently in progress.