New match updating logic and some refactoring
Instead of generating the page every minute, now we just update the
score of each team. The updatePage func runs every minute and gets
the new data from the server. Also now we work with the whole match
row instead of the progress bar and we moved the match datetime to be
a data attribute for the entire row. Because of this it is now easier
to hide or show the progress bar whenever the match is in progress or
not.