New function to check whether the match is on
Some code refactoring to make it easier for reading and understanding.
Extracting a new function that checks whether the match is currently
in progress. That is determined by subtracting the datetime for the match
with the current datetime. If the result is larger than 0 then the match
has begun. The match ends when the passed minutes are more than the
average match length of our choice.